test(ulw-loop): split cli command behavior tests
